# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "~HackPilot, A-Z hackathon teams manager",
        "description": "Automatically generated API documentation  ~HackPilot",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": ["https", "http"],
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "example": "srinivasbatthula.mypc@gmail.com"
                                },
                                "password": {
                                    "example": "1234"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/auth/google/callback": {
            "get": {
                "description": "",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "501": {
                        "description": "Not Implemented"
                    }
                }
            }
        },
...
    }
}